Refactoring:
Eliminate dal modello delle entit?, tutte le classi per gestire le Correzioni.
Sono tornato alla ipotesi di fare una normale collezioni di oggetti e serializzarla come stringa xml
in un campo del database diretto.

Eliminato il Provinatore in favore di una Correzione di tipo Resize.

Eliminate le classi "Comando" e derivate. Ora faccio tutto dentro il servizio apposito IFotoRitoccoSrv.

Creata nuova schermata per il fotoritocco e implementato il Bianco-Nero, Ruota, Torna-Originale
  (piccolo bug: le correzioni non vengono correttamente scritte nel db)

+++ Digiphoto.Lumen.Core/src/Imaging/Immagine.cs
@@ -21,11 +21,11 @@ namespace Digiphoto.Lumen.Imaging {
 
 
 		#region Metodi astratti dall'interfaccia
-		public abstract int ww {
+		public abstract long ww {
 			get;
 		}
 
-		public abstract int hh {
+		public abstract long hh {
 			get;
 		}
